aspire_aligner/requirements.txt
langdetect==1.0.8
xlwt==1.3.0
nltk==3.5
pandas==1.2.0
python_Levenshtein==0.12.0
scikit_learn==0.24.0
numpy==1.19.3
xlrd==2.0.1